Rewrite manga details to be mostly ssr
