+++ src/lib/malSync.ts
@@ -1,6 +1,7 @@
 import db from "@/lib/db";
 import { MalSync } from "@/app/api/interfaces";
 import { getSetting } from "./settings";
+import { getProductionUrl } from "@/app/api/baseUrl";
 
 async function getHqData(malSyncData: MalSync) {
     let service;
@@ -13,7 +14,9 @@ async function getHqData(malSyncData: MalSync) {
         id = malSyncData.aniId;
     } else return null;
 
-    const response = await fetch(`/api/manga/${service}/${id}`);
+    const response = await fetch(
+        `${getProductionUrl()}/api/manga/${service}/${id}`,
+    );
     if (!response.ok) {
         return null;
     }
@@ -26,10 +29,13 @@ export async function fetchMalData(
     overWrite: boolean = false,
     retryCount: number = 3, // Maximum number of retries
     retryDelay: number = 2000, // Delay between retries in milliseconds (2 seconds)
+    useCache: boolean = true,
 ) {
-    const cachedManga = await db.getCache(db.hqMangaCache, identifier);
-    if (cachedManga && cachedManga.score) {
-        return cachedManga;
+    if (useCache) {
+        const cachedManga = await db.getCache(db.hqMangaCache, identifier);
+        if (cachedManga && cachedManga.score) {
+            return cachedManga;
+        }
     }
 
     if (!getSetting("fetchMalImage") && !overWrite) {
@@ -72,7 +78,9 @@ export async function fetchMalData(
                 data["aniUrl"] = malSyncResponseData.aniUrl;
             }
 
-            await db.updateCache(db.hqMangaCache, identifier, data);
+            if (useCache) {
+                await db.updateCache(db.hqMangaCache, identifier, data);
+            }
             return data;
         } catch (error) {
             console.error(error);
